requirements.txt
numpy==1.23.1
pillow==9.2.0
